Systems-Programming-Assignment-2
This file contains the main function of a program that reads from either
a text or binary file and converts to the other form. The Program can
also read a binary file and
determine some statistics. The input and
output files are specified by command line arguments as is the mode.